Refactor: older version message module is renamed to messageOld
