+++ backend/src/main/java/peer/backend/controller/messageOld/MessageController.java
@@ -1,35 +1,36 @@
-package peer.backend.repository.message;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+package peer.backend.controller.messageOld;
 
 import java.time.LocalDate;
-import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import peer.backend.entity.message.Message;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import peer.backend.entity.messageOld.Message;
+import peer.backend.dto.message.MessageSendRequest;
 import peer.backend.entity.user.User;
+import peer.backend.repository.messageOld.MessageRepository;
 import peer.backend.repository.user.UserRepository;
+import peer.backend.service.messageOld.MessageService;
 
-@DisplayName("Message Repository 테스트")
-@DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class MessageRepositoryTest {
-    @Autowired
-    MessageRepository messageRepository;
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/v1/api/message")
+public class MessageController {
 
-    @Autowired
-    UserRepository userRepository;
+    private final MessageService messageService;
 
-    User user0, user1, user2, user3;
+    private final MessageRepository messageRepository;
+    private final UserRepository userRepository;
 
-    @BeforeEach
-    void setting() {
-        user0 = User.builder()
+    @GetMapping("/setting")
+    public void setting()
+    {
+        User user0 = User.builder()
             .userId("userId123")
             .password("password")
             .name("John")
@@ -46,7 +47,7 @@ class MessageRepositoryTest {
             .representAchievement("Achievement XYZ")
             .build();
 
-        user1 = User.builder()
+        User user1 = User.builder()
             .userId("user1")
             .password("password1")
             .name("User One")
@@ -63,7 +64,7 @@ class MessageRepositoryTest {
             .representAchievement("Achievement ABC")
             .build();
 
-        user2 = User.builder()
+        User user2 = User.builder()
             .userId("user2")
             .password("password2")
             .name("User Two")
@@ -80,7 +81,7 @@ class MessageRepositoryTest {
             .representAchievement("Achievement DEF")
             .build();
 
-        user3 = User.builder()
+        User user3 = User.builder()
             .userId("user3")
             .password("password3")
             .name("User Three")
@@ -102,17 +103,6 @@ class MessageRepositoryTest {
         userRepository.save(user2);
         userRepository.save(user3);
 
-    }
-
-//    @Test
-//    public void userCountTest()
-//    {
-//        assertThat(userRepository.count()).isEqualTo(4);
-//    }
-
-    @Test
-    public void findBySenderOrReceiverTest()
-    {
         Message message1 = Message.builder()
             .content("예시1")
             .sender(user0)
@@ -120,31 +110,31 @@ class MessageRepositoryTest {
             .build();
 
         Message message2 = Message.builder()
-            .content("예시1")
+            .content("예시2")
             .sender(user0)
             .receiver(user1)
             .build();
 
         Message message3 = Message.builder()
-            .content("예시1")
+            .content("예시3")
             .sender(user1)
             .receiver(user2)
             .build();
 
         Message message4 = Message.builder()
-            .content("예시1")
+            .content("예시4")
             .sender(user1)
             .receiver(user2)
             .build();
 
         Message message5 = Message.builder()
-            .content("예시1")
+            .content("예시5")
             .sender(user2)
             .receiver(user1)
             .build();
 
         Message message6 = Message.builder()
-            .content("예시1")
+            .content("예시6")
             .sender(user2)
             .receiver(user0)
             .build();
@@ -155,14 +145,23 @@ class MessageRepositoryTest {
         messageRepository.save(message4);
         messageRepository.save(message5);
         messageRepository.save(message6);
-//        assertThat(messageRepository.count()).isEqualTo(6);
-        List<Message> messages = messageRepository.findBySenderOrReceiver(user0, user0);
-        for (Message message : messages)
-        {
-            System.out.println("id = " + message.getId());
-            System.out.println("send = " + message.getSender().getNickname());
-            System.out.println("rec = " + message.getReceiver().getNickname());
-        }
-        assertThat(messageRepository.findBySenderOrReceiver(user0, user0).size()).isEqualTo(3);
+    }
+
+    @GetMapping("/list/{userId}")
+    public ResponseEntity userMessageList(@PathVariable("userId") Long userId)
+    {
+        return new ResponseEntity(messageService.myMessageList(userId), HttpStatus.OK);
+    }
+
+    @GetMapping("/{userId}")
+    public ResponseEntity messageDetail(@PathVariable("userId") Long userId)
+    {
+        return new ResponseEntity(messageService.userDetailMessage(userId), HttpStatus.OK);
+    }
+
+    @PostMapping("/{userId}")
+    public ResponseEntity sendMessage(@PathVariable("userId") Long userId,
+        @RequestBody MessageSendRequest messageSendRequest) {
+        return new ResponseEntity(messageService.sendMessage(userId, messageSendRequest), HttpStatus.OK);
     }
 }
