+++ backend/src/main/java/peer/backend/dto/message/MessageDTO.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import peer.backend.entity.message.MessageType;
+import peer.backend.entity.messageOld.MessageType;
 
 @Getter
 @Builder
