+++ backend/src/main/java/peer/backend/repository/messageOld/MessageRepository.java
@@ -1,8 +1,8 @@
-package peer.backend.repository.message;
+package peer.backend.repository.messageOld;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import peer.backend.entity.message.Message;
+import peer.backend.entity.messageOld.Message;
 import peer.backend.entity.user.User;
 
 public interface MessageRepository extends JpaRepository<Message, Long> {
